+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 - Removes `include` entry from manifest as `exclude` can provide same functionality now that it follows .gitignore semantics (such as `!` symbol)
 - Renames `ORBIT_MANIFEST_PATH` environment variable to `ORBIT_MANIFEST_FILE`
 - Renames `ORBIT_CHAN_INDEX` environment variable to `ORBIT_CHANNEL_IP_DIR`
+- Extends name mangling during dynamic symbol transformation from 10 characters of the checksum to 16 characters
 
 ### Fixes
 - Improves error message when giving bad version for local/relative dependency